requirements.txt
sqlalchemy>=1.4
psycopg2-binary
fastapi
uvicorn
passlib
argon2-cffi
python-jose
